requirements.txt
live2d-py
PyOpenGL>=3.1.0
PyAudio>=0.2.13
PyQt6==6.6.1
ollama
RealtimeSTT>=0.1.6